Add native NeoForge support
- Added native NeoForge support (closes #22)
- Fixed Forge version crashing in 1.16x-1.17x with Java 8 or Java 16 due to Init trying to load Java 17 bytecode before InitLegacy
