+++ META-INF/neoforge.mods.toml
@@ -1,0 +1,43 @@
+modLoader = "javafml"
+loaderVersion="[3,)"
+license="Custom"
+logoFile="pack.png"
+logoBlur=false
+showAsDataPack=true
+
+[[mods]]
+modId="moremobs"
+version="${version}+mod"
+displayName="More Mobs"
+description='''Humanoid mobs get new variants through custom heads or through unused vanilla mobs! This pack uses §lover 85 custom player heads§r which can be obtained and placed in your world.
+No Resource pack needed!
+
+§lSupported Entities:§r Zombies, Husks, Drowned, Skeletons, Strays, Wither Skeletons, Mooshrooms, Piglins, Pillagers, Evokers and Vindicators
+
+Execute the command §o/function §omore_mobs:settings§r for a world-specific settings menu and/or install MidnightLib for global settings across all worlds.
+'''
+authors="Tschipcraft"
+displayURL="https://modrinth.com/mod/HJR6V0I2"
+issueTrackerURL="https://github.com/Tschipcraft/more_mobs/issues"
+updateJSONURL="https://api.modrinth.com/updates/HJR6V0I2/forge_updates.json?neoforge=only"
+displayTest="IGNORE_SERVER_VERSION"
+features={ java_version = "[21,)" }
+
+[[dependencies.moremobs]]
+    modId="minecraft"
+    mandatory=true
+    versionRange="${mc_version_range_neoforge}"
+    ordering="NONE"
+    side="BOTH"
+
+[[dependencies.moremobs]]
+    modId="midnightlib"
+    type="optional"
+    versionRange="[1.0.0,)"
+    ordering="NONE"
+    side="BOTH"
+    referralUrl="https://modrinth.com/mod/codAaoxh/versions?l=neoforge"
+    reason="Global settings across all worlds"
+
+[modproperties.moremobs]
+    catalogueItemIcon="pack.png"
