+++ data/more_mobs/functions/install.mcfunction
@@ -19,6 +19,8 @@ scoreboard objectives add ts.mm.nt_test dummy
 # Menu
 scoreboard objectives add ts.mm.welcome dummy
 scoreboard objectives add tschipcraft.menu trigger
+scoreboard players reset @a tschipcraft.menu
+scoreboard players enable @a tschipcraft.menu
 
 scoreboard objectives add tvc_ignore dummy
