+++ src/main/java/net/tschipcraft/moremobs/fabric/sendConfig.java
@@ -70,4 +70,5 @@ public class sendConfig {
         } catch (CommandSyntaxException ignored) {
         }
     }
+
 }
