+++ src/main/java/net/tschipcraft/moremobs/forge/Init.java
@@ -5,18 +5,12 @@ import eu.midnightdust.lib.config.MidnightConfig;
 import net.minecraftforge.client.ConfigScreenHandler;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.server.ServerStartedEvent;
-import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.ModLoadingContext;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.slf4j.Logger;
 
-// This is the actual init class. It gets translated into Java 17
-@Mod("moremobs")
-@Mod.EventBusSubscriber(modid = "moremobs")
+// This class handles registering the MidnightLib config. It gets translated into Java 17 bytecode.
 public class Init {
 
 	public static final String MODID = "moremobs";
@@ -27,12 +21,13 @@ public class Init {
 			// Directly reference a slf4j logger
 			LOGGER = LogUtils.getLogger();
 		} catch (NoClassDefFoundError ignored) {
+			// No logging
 		}
 	}
 
 	public static void registerEvent() {
 		try {
-			if (LOGGER != null) LOGGER.info("[More Mobs] Registering server started event...");
+			if (LOGGER != null) LOGGER.info("[More Mobs] Registering server started event ...");
 			Class.forName("net.minecraftforge.event.server.ServerStartedEvent");
 
 			// Register onServerStarted handler
@@ -40,13 +35,14 @@ public class Init {
 				@SubscribeEvent
 				public void onServerStarted(ServerStartedEvent event) {
 					if (ModList.get().isLoaded("midnightlib")) {
-						LOGGER.info("[More Mobs] Sending global config to world...");
+						LOGGER.info("[More Mobs] Sending global config to world ...");
 						sendConfig.sendConfig(event.getServer());
 					}
 				}
 			});
 		} catch (ClassNotFoundException e) {
 			// ServerStartedEvent doesn't exist, don't register
+			if (LOGGER != null) LOGGER.info("[More Mobs] ServerStartedEvent not found, not registering event handler.");
 		}
 	}
 
@@ -65,45 +61,4 @@ public class Init {
 		}
 	}
 
-	public Init() {
-		IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
-
-		// Register the setup method for modloading
-		modEventBus.addListener(this::setup);
-
-		try {
-			if (LOGGER != null) LOGGER.info("[More Mobs] Registering server started event...");
-			Class.forName("net.minecraftforge.event.server.ServerStartedEvent");
-
-			// Register onServerStarted handler
-			MinecraftForge.EVENT_BUS.register(new Object() {
-				@SubscribeEvent
-				public void onServerStarted(ServerStartedEvent event) {
-					if (ModList.get().isLoaded("midnightlib")) {
-						LOGGER.info("[More Mobs] Sending global config to world...");
-						sendConfig.sendConfig(event.getServer());
-					}
-				}
-			});
-
-		} catch (ClassNotFoundException e) {
-			// ServerStartedEvent doesn't exist, don't register
-		}
-	}
-
-	private void setup(final FMLCommonSetupEvent event) {
-		if (ModList.get().isLoaded("midnightlib")) {
-			// Use MidnightLib features
-			LOGGER.info("[More Mobs] MidnightLib detected!");
-			// Build config class
-			Config.init(MODID, Config.class);
-			// Initialize config screen
-			ModLoadingContext.get().registerExtensionPoint(ConfigScreenHandler.ConfigScreenFactory.class, () -> {
-				return new ConfigScreenHandler.ConfigScreenFactory((client, parent) -> {
-					return MidnightConfig.getScreen(parent, MODID);
-				});
-			});
-		}
-		if (LOGGER != null) LOGGER.info("[More Mobs] Loaded More Mobs by Tschipcraft successfully!");
-	}
 }
